+++ plugin/src/test/java/io/jenkins/plugins/casc/vault/VaultSecretSourceTest.java
@@ -49,7 +49,7 @@ public class VaultSecretSourceTest {
     public void kv1WithUser() {
         envVars.set("CASC_VAULT_USER", VAULT_USER);
         envVars.set("CASC_VAULT_PW", VAULT_PW);
-        envVars.set("CASC_VAULT_PATH", VAULT_PATH_V1);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV1_1 + "," + VAULT_PATH_KV1_2);
         envVars.set("CASC_VAULT_ENGINE_VERSION", "1");
         assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo("123"));
     }
@@ -58,7 +58,7 @@ public class VaultSecretSourceTest {
     public void kv2WithUser() {
         envVars.set("CASC_VAULT_USER", VAULT_USER);
         envVars.set("CASC_VAULT_PW", VAULT_PW);
-        envVars.set("CASC_VAULT_PATH", VAULT_PATH_V2);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV2_1 + "," + VAULT_PATH_KV2_2);
         envVars.set("CASC_VAULT_ENGINE_VERSION", "2");
         assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo("123"));
     }
@@ -67,7 +67,7 @@ public class VaultSecretSourceTest {
     public void kv2WithWrongUser() {
         envVars.set("CASC_VAULT_USER", "1234");
         envVars.set("CASC_VAULT_PW", VAULT_PW);
-        envVars.set("CASC_VAULT_PATH", VAULT_PATH_V2);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV2_1);
         envVars.set("CASC_VAULT_ENGINE_VERSION", "2");
         assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo(""));
     }
@@ -75,7 +75,7 @@ public class VaultSecretSourceTest {
     @Test
     public void kv1WithToken() {
         envVars.set("CASC_VAULT_TOKEN", VAULT_ROOT_TOKEN);
-        envVars.set("CASC_VAULT_PATH", VAULT_PATH_V1);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV1_1 + "," + VAULT_PATH_KV1_2);
         envVars.set("CASC_VAULT_ENGINE_VERSION", "1");
         assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo("123"));
     }
@@ -83,7 +83,7 @@ public class VaultSecretSourceTest {
     @Test
     public void kv2WithToken() {
         envVars.set("CASC_VAULT_TOKEN", VAULT_ROOT_TOKEN);
-        envVars.set("CASC_VAULT_PATH", VAULT_PATH_V2);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV2_1 + "," + VAULT_PATH_KV2_2);
         envVars.set("CASC_VAULT_ENGINE_VERSION", "2");
         assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo("123"));
     }
@@ -91,7 +91,7 @@ public class VaultSecretSourceTest {
     @Test
     public void kv1WithWrongToken() {
         envVars.set("CASC_VAULT_TOKEN", "1234");
-        envVars.set("CASC_VAULT_PATH", VAULT_PATH_V1);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV1_1);
         envVars.set("CASC_VAULT_ENGINE_VERSION", "1");
         assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo(""));
     }
@@ -100,7 +100,7 @@ public class VaultSecretSourceTest {
     public void kv1WithApprole() {
         envVars.set("CASC_VAULT_APPROLE", VAULT_APPROLE_ID);
         envVars.set("CASC_VAULT_APPROLE_SECRET", VAULT_APPROLE_SECRET);
-        envVars.set("CASC_VAULT_PATH", VAULT_PATH_V1);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV1_1 + "," + VAULT_PATH_KV1_2);
         envVars.set("CASC_VAULT_ENGINE_VERSION", "1");
         assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo("123"));
     }
@@ -109,7 +109,7 @@ public class VaultSecretSourceTest {
     public void kv2WithApprole() {
         envVars.set("CASC_VAULT_APPROLE", VAULT_APPROLE_ID);
         envVars.set("CASC_VAULT_APPROLE_SECRET", VAULT_APPROLE_SECRET);
-        envVars.set("CASC_VAULT_PATH", VAULT_PATH_V2);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV2_1 + "," + VAULT_PATH_KV2_2);
         envVars.set("CASC_VAULT_ENGINE_VERSION", "2");
         assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo("123"));
     }
@@ -118,8 +118,38 @@ public class VaultSecretSourceTest {
     public void kv2WithWrongApprole() {
         envVars.set("CASC_VAULT_APPROLE", "1234");
         envVars.set("CASC_VAULT_APPROLE_SECRET", VAULT_APPROLE_SECRET);
-        envVars.set("CASC_VAULT_PATH", VAULT_PATH_V2);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV2_1);
         envVars.set("CASC_VAULT_ENGINE_VERSION", "2");
         assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo(""));
+    }
+
+    @Test
+    public void kv2WithApproleMultipleKeys() {
+        envVars.set("CASC_VAULT_APPROLE", VAULT_APPROLE_ID);
+        envVars.set("CASC_VAULT_APPROLE_SECRET", VAULT_APPROLE_SECRET);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV2_1 + "," + VAULT_PATH_KV2_2);
+        envVars.set("CASC_VAULT_ENGINE_VERSION", "2");
+        assertThat(SecretSourceResolver.resolve(context, "${key2}"), equalTo("456"));
+        assertThat(SecretSourceResolver.resolve(context, "${key3}"), equalTo("789"));
+    }
+
+    @Test
+    public void kv2WithApproleMultipleKeysOverriden() {
+        envVars.set("CASC_VAULT_APPROLE", VAULT_APPROLE_ID);
+        envVars.set("CASC_VAULT_APPROLE_SECRET", VAULT_APPROLE_SECRET);
+        envVars.set("CASC_VAULT_PATHS", VAULT_PATH_KV2_1 + "," + VAULT_PATH_KV2_2 + "," + VAULT_PATH_KV2_3);
+        envVars.set("CASC_VAULT_ENGINE_VERSION", "2");
+        assertThat(SecretSourceResolver.resolve(context, "${key2}"), equalTo("321"));
+        assertThat(SecretSourceResolver.resolve(context, "${key1}"), equalTo("123"));
+    }
+
+    // TODO: used to check for backwards compatibility. Deprecate!
+    @Test
+    public void kv2WithUserDeprecatedPath() {
+        envVars.set("CASC_VAULT_APPROLE", VAULT_APPROLE_ID);
+        envVars.set("CASC_VAULT_APPROLE_SECRET", VAULT_APPROLE_SECRET);
+        envVars.set("CASC_VAULT_PATH", VAULT_PATH_KV1_1 + "," + VAULT_PATH_KV1_2);
+        envVars.set("CASC_VAULT_ENGINE_VERSION", "1");
+        assertThat(SecretSourceResolver.resolve(context, "${key3}"), equalTo("789"));
     }
 }